+++ include/managers/ble_manager.h
@@ -43,5 +43,11 @@ void ble_start_scanning(void);
 void ble_start_skimmer_detection(void);
 void ble_stop_skimmer_detection(void);
 
+// AirTag specific functions
+void ble_list_airtags(void);
+void ble_select_airtag(int index);
+void ble_start_spoofing_selected_airtag(void);
+void ble_stop_spoofing(void);
+
 #endif
 #endif // BLE_MANAGER_H