+++ main/managers/views/options_screen.c
@@ -96,6 +96,8 @@ static const char *wifi_options[] = {"Scan Access Points",
                                      NULL};
 
 static const char *bluetooth_options[] = {"Find Flippers",   "Start AirTag Scanner",
+                                          "List AirTags",    "Select AirTag",
+                                          "Spoof Selected AirTag", "Stop Spoofing",
                                           "Raw BLE Scanner", "BLE Skimmer Detect",
                                           "Go Back",         NULL};
 
@@ -537,6 +539,50 @@ void option_event_cb(lv_event_t *e) {
 #ifndef CONFIG_IDF_TARGET_ESP32S2
         display_manager_switch_view(&terminal_view);
         simulateCommand("blescan -f");
+        view_switched = true;
+#else
+        error_popup_create("Device Does not Support Bluetooth...");
+        
+#endif
+    }
+
+    else if (strcmp(Selected_Option, "List AirTags") == 0) {
+#ifndef CONFIG_IDF_TARGET_ESP32S2
+        display_manager_switch_view(&terminal_view);
+        simulateCommand("listairtags");
+        view_switched = true;
+#else
+        error_popup_create("Device Does not Support Bluetooth...");
+        
+#endif
+    }
+
+    else if (strcmp(Selected_Option, "Select AirTag") == 0) {
+#ifndef CONFIG_IDF_TARGET_ESP32S2
+        set_number_pad_mode(NP_MODE_AIRTAG);
+        display_manager_switch_view(&number_pad_view);
+        view_switched = true;
+#else
+        error_popup_create("Device Does not Support Bluetooth...");
+        
+#endif
+    }
+
+    else if (strcmp(Selected_Option, "Spoof Selected AirTag") == 0) {
+#ifndef CONFIG_IDF_TARGET_ESP32S2
+        display_manager_switch_view(&terminal_view);
+        simulateCommand("spoofairtag");
+        view_switched = true;
+#else
+        error_popup_create("Device Does not Support Bluetooth...");
+        
+#endif
+    }
+
+    else if (strcmp(Selected_Option, "Stop Spoofing") == 0) {
+#ifndef CONFIG_IDF_TARGET_ESP32S2
+        display_manager_switch_view(&terminal_view);
+        simulateCommand("stopspoof");
         view_switched = true;
 #else
         error_popup_create("Device Does not Support Bluetooth...");